tmdeploy/requirements.txt
ansible==2.3.2.0
ansible-container==0.3.0
apache-libcloud==2.1.0
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.5.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.3
boto3==1.4.6
botocore==1.6.8
cached-property==1.3.0
cffi==1.10.0
cliff==2.8.0
cmd2==0.7.7
cryptography==2.0.3
debtcollector==1.17.0
decorator==4.1.2
deprecation==1.0.1
docker-compose==1.9.0
docker-py==1.10.6
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dogpile.cache==0.6.4
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
idna==2.6
ipaddress==1.0.18
iso8601==0.1.12
Jinja2==2.9.6
jmespath==0.9.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==3.1.0
MarkupSafe==1.0
monotonic==1.3
msgpack-python==0.4.8
munch==2.2.0
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.9.17
os-client-config==1.28.0
osc-lib==1.7.0
oslo.config==4.11.0
oslo.i18n==3.17.0
oslo.serialization==2.20.0
oslo.utils==3.28.0
paramiko==2.2.3
pbr==3.1.1
positional==1.2.1
prettytable==0.7.2
psycopg2==2.7.3.1
pyasn1==0.3.3
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.1.2
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyperclip==1.5.27
python-cinderclient==3.1.0
python-dateutil==2.6.1
python-glanceclient==2.8.0
python-ironicclient==1.16.0
python-keystoneclient==3.13.0
python-novaclient==9.1.0
python-openstackclient==3.12.0
pytz==2017.2
PyYAML==3.12
requests==2.11.1
requestsexceptions==1.3.0
rfc3986==1.1.0
s3transfer==0.1.10
shade==1.22.2
simplejson==3.11.1
six==1.10.0
stevedore==1.25.0
texttable==0.8.8
tmdeploy
unicodecsv==0.14.1
warlock==1.2.0
websocket-client==0.44.0
whichcraft==0.4.1
wrapt==1.10.11
